DFE-666: Breakout view models from service
