+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/Services/Interfaces/IContentService.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using GovUk.Education.ExploreEducationStatistics.Content.Api.ViewModels;
 
 namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Services.Interfaces
 {
