+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/ViewModels/PublicationTree.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace GovUk.Education.ExploreEducationStatistics.Content.Api.ViewModels
+{
+    public class PublicationTree
+
+    {
+        public Guid Id { get; set; }
+        public string Title { get; set; }
+        public string Slug { get; set; }
+        public string Summary { get; set; }
+    }
+}
