+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/ViewModels/ThemeTree.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+
+namespace GovUk.Education.ExploreEducationStatistics.Content.Api.ViewModels
+{
+    public class ThemeTree
+    {
+        public Guid Id { get; set; }
+        public string Title { get; set; }
+
+        public List<TopicTree> Topics { get; set; }
+    }
+}
